Web_demo/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
Pillow>=9.5.0
numpy>=1.24.0